+++ android/src/main/java/com/twiliovoicereactnative/VoiceNotificationReceiver.java
@@ -125,8 +125,7 @@ public class VoiceNotificationReceiver extends BroadcastReceiver {
     Notification notification = NotificationUtility.createIncomingCallNotification(
       context.getApplicationContext(),
       callRecord,
-      VOICE_CHANNEL_HIGH_IMPORTANCE,
-      true);
+      VOICE_CHANNEL_HIGH_IMPORTANCE);
     createOrReplaceNotification(context, callRecord.getNotificationId(), notification);
 
     // play ringer sound
@@ -282,8 +281,7 @@ public class VoiceNotificationReceiver extends BroadcastReceiver {
     Notification notification = NotificationUtility.createIncomingCallNotification(
       context.getApplicationContext(),
       callRecord,
-      VOICE_CHANNEL_DEFAULT_IMPORTANCE,
-      false);
+      VOICE_CHANNEL_DEFAULT_IMPORTANCE);
     createOrReplaceNotification(context, callRecord.getNotificationId(), notification);
 
     // stop active sound (if any)