+++ src/__tests__/CallInvite.test.ts
@@ -7,7 +7,7 @@ import type { NativeEventEmitter as MockNativeEventEmitterType } from '../__mock
 import { Call } from '../Call';
 import { CallInvite } from '../CallInvite';
 import { TwilioError } from '../error/TwilioError';
-import { NativeEventEmitter, NativeModule } from '../common';
+import { NativeEventEmitter, NativeModule, Platform } from '../common';
 import { Constants } from '../constants';
 import type { NativeCallInviteEvent } from '../type/CallInvite';
 
@@ -394,6 +394,48 @@ describe('CallInvite class', () => {
       ).getTo();
       expect(typeof to).toBe('string');
       expect(to).toBe(createNativeCallInviteInfo().to);
+    });
+  });
+
+  describe('.updateCallerHandle()', () => {
+    it('should resolve on ios platforms', async () => {
+      const platformSpy = jest
+        .spyOn(Platform, 'OS', 'get')
+        .mockReturnValue('ios');
+      const nativeMethodSpy = jest.spyOn(
+        NativeModule,
+        'callInvite_updateCallerHandle'
+      );
+      await expect(
+        new CallInvite(
+          createNativeCallInviteInfo(),
+          CallInvite.State.Pending
+        ).updateCallerHandle('foobar')
+      ).resolves.toBeUndefined();
+      expect(platformSpy.mock.calls).toEqual([[]]);
+      expect(nativeMethodSpy.mock.calls).toEqual([
+        ['mock-nativecallinviteinfo-uuid', 'foobar'],
+      ]);
+    });
+
+    it('should reject on android platforms', async () => {
+      const platformSpy = jest
+        .spyOn(Platform, 'OS', 'get')
+        .mockReturnValue('android');
+      const nativeMethodSpy = jest.spyOn(
+        NativeModule,
+        'callInvite_updateCallerHandle'
+      );
+      await expect(
+        new CallInvite(
+          createNativeCallInviteInfo(),
+          CallInvite.State.Pending
+        ).updateCallerHandle('foobar')
+      ).rejects.toThrowError(
+        'Unsupported platform "android". This method is only supported on iOS.'
+      );
+      expect(platformSpy.mock.calls).toEqual([[], []]);
+      expect(nativeMethodSpy.mock.calls).toEqual([]);
     });
   });
 });