+++ CHANGELOG.md
@@ -35,9 +35,12 @@ Twilio Voice React Native SDK has now reached milestone `beta.5`. Included in th
 
 #### Android
 - Call timestamp now in simplified ISO-8601 format, not stored as a double from epoch.
+- Uses system provided notification styles for incoming & ongoing calls. This insures visual consistency between devices.
+
 
 #### iOS
 - The call connected timestamp is now in simplified ISO-8601 format.
+- A new method `CallInvite.updateCallerHandle()` has been added. Use this method to update the caller's name displayed in the iOS system incoming call UI. This method is specific to iOS and unavailable in Android.
 
 1.0.0-beta.4 (Jan 11, 2024)
 ===========================